backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-dotenv
nltk
requests
openai>=1.0.0